+++ src/rpm/signature/traits.rs
@@ -1,19 +1,9 @@
 //! Trait abstractions of signing operations.
-//!
-//! Does not contain hashing! Hashes are fixed by the rpm
-//! "spec" to sha1, md5 (yes, that is correct), sha2_256.
 
 use crate::Timestamp;
 use crate::errors::*;
 use std::fmt::Debug;
 use std::io;
-
-#[derive(Clone, Copy, Debug)]
-pub enum AlgorithmType {
-    RSA,
-    ECDSA,
-    EdDSA,
-}
 
 /// Signing trait to be implement for RPM signing.
 pub trait Signing: Debug
@@ -22,7 +12,6 @@ where
 {
     type Signature;
     fn sign(&self, data: impl io::Read, t: Timestamp) -> Result<Self::Signature, Error>;
-    fn algorithm(&self) -> AlgorithmType;
 }
 
 impl<T, S> Signing for &T
@@ -34,10 +23,6 @@ where
     fn sign(&self, data: impl io::Read, t: Timestamp) -> Result<Self::Signature, Error> {
         T::sign(self, data, t)
     }
-
-    fn algorithm(&self) -> AlgorithmType {
-        T::algorithm(self)
-    }
 }
 
 /// Verification trait to be implement for RPM signature verification.
@@ -47,7 +32,6 @@ where
 {
     type Signature;
     fn verify(&self, data: impl io::Read, signature: &[u8]) -> Result<(), Error>;
-    fn algorithm(&self) -> AlgorithmType;
 }
 
 impl<T, S> Verifying for &T
@@ -59,10 +43,6 @@ where
     fn verify(&self, data: impl io::Read, signature: &[u8]) -> Result<(), Error> {
         T::verify(self, data, signature)
     }
-
-    fn algorithm(&self) -> AlgorithmType {
-        T::algorithm(self)
-    }
 }
 
 pub mod key {
@@ -70,8 +50,7 @@ pub mod key {
     /// Marker trait for key types.
     pub trait KeyType: super::Debug + Copy {}
 
-    /// A secret key that should not be shared with any other party
-    /// under any circumstance.
+    /// A secret key that should not be shared with any other party under any circumstance.
     #[derive(Debug, Clone, Copy)]
     pub struct Secret;
 
@@ -89,10 +68,6 @@ impl<T> Signing for std::marker::PhantomData<T> {
     fn sign(&self, _data: impl io::Read, _t: Timestamp) -> Result<Self::Signature, Error> {
         unreachable!("you need to implement `sign` of the `Signing` trait")
     }
-
-    fn algorithm(&self) -> AlgorithmType {
-        unreachable!("you need to implement `algorithm` of the `Signing` trait")
-    }
 }
 
 /// Implement unreachable verifier for the empty tuple`()`
@@ -100,9 +75,5 @@ impl<T> Verifying for std::marker::PhantomData<T> {
     type Signature = Vec<u8>;
     fn verify(&self, _data: impl io::Read, _x: &[u8]) -> Result<(), Error> {
         unreachable!("you need to implement `verify` of the `Verifying` trait")
-    }
-
-    fn algorithm(&self) -> AlgorithmType {
-        unreachable!("you need to implement `algorithm` of the `Verifying` trait")
     }
 }